essage
git-svn-id: https://vvvapp.svn.sourceforge.net/svnroot/vvvapp/trunk@59 b0e3558a-1426-0410-83de-bf875247a1ac

+++ src/data_interface/fb_paths.cpp
@@ -118,7 +118,7 @@ wxString CPaths::FB_GetFullPath( long PathID ) {
 	Statement st = StatementFactory( db->GetIBPPDB(), db->TransactionGetReference() );
 
 	st->Prepare( "EXECUTE PROCEDURE SP_GET_FULL_PATH( ?, ? )" );
-	st->Set( 1, PathID );
+	st->Set( 1, (int32_t) PathID );
 	st->Set( 2, wxString(wxFileName::GetPathSeparator()).c_str() );
 	st->Execute();
